fix: markdown embedded html & agent intro not firing onChange
