+++ ui/admin/package.json
@@ -63,6 +63,8 @@
 		"react-resizable-panels": "^2.1.3",
 		"react-router": "^7.0.0",
 		"rehype-external-links": "^3.0.0",
+		"rehype-raw": "^7.0.0",
+		"rehype-sanitize": "^6.0.0",
 		"remark-gfm": "^4.0.0",
 		"safe-routes": "^1.0.2",
 		"sonner": "1.4.3",
